refactor: update resource fields to use i64 instead of i32
- Changed resource-related fields across multiple domain structures, schemas, and migrations from `i32` to `i64`.
- Updated types for `Deduction` and other resource-related structs to support larger numeric ranges.

+++ migrations/2024-03-02-054020_building_levels/up.sql
@@ -4,10 +4,10 @@ CREATE TABLE building_level
     building_id  INTEGER     NOT NULL,
     level        INTEGER     NOT NULL,
     upgrade_time TEXT        NOT NULL, -- "%H:%M:%S"
-    req_food     INTEGER,
-    req_wood     INTEGER,
-    req_stone    INTEGER,
-    req_gold     INTEGER,
+    req_food     BIGINT,
+    req_wood     BIGINT,
+    req_stone    BIGINT,
+    req_gold     BIGINT,
     created_at   TIMESTAMPTZ NOT NULL DEFAULT now(),
     updated_at   TIMESTAMPTZ NOT NULL DEFAULT now(),
